+++ lua/macime/types.lua
@@ -1,4 +1,12 @@
 ---macime.Config
+---@class macime.Config
+---@field vim? macime.Config.Vim
+---@field ime? macime.Config.Ime
+---@field save? macime.Config.Save
+---@field socket? macime.Config.Socket
+---@field include? macime.Config.Include
+---@field exclude? macime.Config.Exclude
+
 ---@class macime.Config.Vim
 ---@field ttimeoutlen? number
 
@@ -20,13 +28,25 @@
 ---@class macime.Config.Exclude
 ---@field filetype? string[]
 
----@class macime.Config
----@field vim? macime.Config.Vim
----@field ime? macime.Config.Ime
----@field save? macime.Config.Save
----@field socket? macime.Config.Socket
----@field include? macime.Config.Include
----@field exclude? macime.Config.Exclude
+---macime.Context
+---@class macime.Context
+---@field macime macime.Context.Macime
+---@field macimed macime.Context.Macimed
+---@field capability macime.Context.Capability
+
+---@class macime.Context.Macime
+---@field installed boolean
+---@field version string
+
+---@class macime.Context.Macimed
+---@field installed boolean
+---@field version string
+---
+---@class macime.Context.Capability
+---@field macime_direct boolean -- `macime` >= 3.2.0
+---@field macimed_daemon boolean -- `macime` >= 3.2.0
+---@field cjk_refresh boolean -- `macime` >= 3.5.0
+---@field daemon_socket_api boolean -- `macime` >= 3.6.0
 
 ---macime.Health
 ---@class macime.Health
@@ -39,8 +59,9 @@
 ---@field macimed_sock_path? string
 ---@field macimed_macime_path? string
 ---@field capability_direct? boolean
----@field capability_socket? boolean
+---@field capability_daemon? boolean
 ---@field capability_cjk_refresh? boolean
+---@field capability_daemon_socket_api? boolean
 ---@field service_name? string
 ---@field service_status? string
 ---@field service_user? string