+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # CHANGELOG.md
 
-## [v2.3.1](https://github.com/riodelphino/macime.nvim/compare/v2.3.0...v2.3.1) (2026-02-04)
+## [v3.0.0](https://github.com/riodelphino/macime.nvim/compare/v2.3.0...v3.0.0) (2026-02-07)
 
+* **feat:** Adapt to `macime` v3.6.0 (Send command like `ime set com.apple...` `daemon get sock-path`)
 * **docs:** Remove `(Fixed value)` from `sock_path` (Now changeable) in checkhealth
 * **docs:** Refine `IME default` in checkhealth
 * **docs:** Add `daemon`(macime path) to `Homebrew Service` in checkhealth